Made the timer start, decrease and stop. Refractored the main.js file
There are now functions to start, tick and pause the timer. All the
functions are in the timer class and they are attached to objects. Also,
events have been created so that when a function is used inside the
timer, it can be communicated to the outside.

The main.js file has been refractored, meaning that the Timer class has
been moved to it's own file - timer.js - and it has been added to the
HTML document.

Next up, the border animation.