Created animations for a circle-border which runs arround the timer
The border that we can see is a an empty circle inside an svg element.
All events are triggered through the callbacks which the instance of the
timer object provides. The CSS and part of the HTML files were added
just for styling purposes and were not explained, but I understand them.