Finished the search and statistics for one movie
The tutorial was very interesting and I forgot to commit.
Learnt a lot about AXIOS and how good it is in practice. I also
understood how to make HTML documents in JS and what the advantages are.
There was also a lot of practice with adding events to elements after
injecting them. Next there is going to be some refractoring.

Right now, the application has a nice dropdown search menu when typing
in the search bar. Clicking on a movie will bring up statistics about it.