An application that enables you to send secrete messages via URL
The application asks the user for the message that they want to send.
Then, the application attaches the text incoded in base64 to the end of
the URL. The URL can then be sent to another user and the message will
be shown when the URL is clicked. Then, the user can click on a link to
create their own secret messages.

The application still needs to be deployed by using the npx now command